@@ -1,8 +1,8 @@
 {
     "name": "codelinered/pdf-image",
-    "description": "Converts uploaded images to PDF and can merge 2 PDF files to 1.",
-    "version": "1.1.0",
-    "keywords": ["zend", "pdf", "image", "convert", "merge"],
+    "description": "Converts uploaded images to PDF, merges 2 PDF files to 1 or converts PDF to image.",
+    "version": "1.2.0-dev",
+    "keywords": ["zend", "pdf", "image", "convert", "merge", "create"],
     "homepage": "https://github.com/CodelineRed/pdf-image",
     "license": "MIT",
     "type": "project",
@@ -13,15 +13,15 @@
             "homepage": "https://www.codelinered.net"
         }
     ],
-    "minimum-stability": "stable",
     "require": {
-        "php": "5.5 - 7.1",
-        "ext-gd": "*"
+        "php": "^8.0",
+        "ext-gd": "*",
+        "ext-imagick": "*"
     },
     "config": {
         "process-timeout": 0,
         "platform": {
-            "php": "7.1.33"
+            "php": "8.2"
         }
     }
 }
