@@ -159,7 +159,7 @@ class Zend_Form_Decorator_Description extends Zend_Form_Decorator_Abstract
         }
 
         $description = $element->getDescription();
-        $description = trim($description);
+        $description = is_string($description) ? trim($description) : null;
 
         if (!empty($description) && (null !== ($translator = $element->getTranslator()))) {
             $description = $translator->translate($description);
