@@ -128,7 +128,7 @@ class Zend_Loader_PluginLoader implements Zend_Loader_PluginLoader_Interface
         }
 
         $last = strlen($prefix) - 1;
-        if ($prefix{$last} == '\\') {
+        if ($prefix[$last] == '\\') {
             return $prefix;
         }
 
