@@ -85,48 +85,48 @@ class Zend_Pdf_NameTree implements ArrayAccess, Iterator, Countable
         }
     }
 
-    public function current()
+    public function current(): mixed
     {
         return current($this->_items);
     }
 
 
-    public function next()
+    public function next(): void
     {
-        return next($this->_items);
+        next($this->_items);
     }
 
 
-    public function key()
+    public function key(): mixed
     {
         return key($this->_items);
     }
 
 
-    public function valid() {
+    public function valid(): bool {
         return current($this->_items)!==false;
     }
 
 
-    public function rewind()
+    public function rewind(): void
     {
         reset($this->_items);
     }
 
 
-    public function offsetExists($offset)
+    public function offsetExists($offset): bool
     {
         return array_key_exists($offset, $this->_items);
     }
 
 
-    public function offsetGet($offset)
+    public function offsetGet($offset): mixed
     {
         return $this->_items[$offset];
     }
 
 
-    public function offsetSet($offset, $value)
+    public function offsetSet($offset, $value): void
     {
         if ($offset === null) {
             $this->_items[]        = $value;
@@ -136,7 +136,7 @@ class Zend_Pdf_NameTree implements ArrayAccess, Iterator, Countable
     }
 
 
-    public function offsetUnset($offset)
+    public function offsetUnset($offset): void
     {
         unset($this->_items[$offset]);
     }
@@ -147,7 +147,7 @@ class Zend_Pdf_NameTree implements ArrayAccess, Iterator, Countable
         $this->_items = array();
     }
 
-    public function count()
+    public function count(): int
     {
         return count($this->_items);
     }
