@@ -514,7 +514,7 @@ class Zend_Pdf_StringParser
             return null;
         }
 
-        $ref = new Zend_Pdf_Element_Reference((int)$objNum, (int)$genNum, $this->_context, $this->_objFactory->resolve());
+        $ref = new Zend_Pdf_Element_Reference((int)$objNum, $this->_context, $this->_objFactory->resolve(), (int)$genNum);
 
         return $ref;
     }
