@@ -122,9 +122,11 @@ class Zend_Validate_File_Extension extends Zend_Validate_Abstract
      */
     public function getExtension()
     {
-        $extension = explode(',', $this->_extension);
+        if (is_string($this->_extension)) {
+            return explode(',', $this->_extension);
+        }
 
-        return $extension;
+        return [];
     }
 
     /**
