@@ -223,7 +223,7 @@ class Zend_Validate_File_Upload extends Zend_Validate_Abstract
             }
         }
 
-        if (count($this->_messages) > 0) {
+        if (is_array($this->_messages) && count($this->_messages) > 0) {
             return false;
         } else {
             return true;
